requirements.txt
simpy>=4.1.1
pandas>=2.0.0
numpy>=1.23.0
plotly>=5.22.0
streamlit>=1.33.0
pyyaml>=6.0.1
